Addressed the review comments
